% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudview-package.R
\docType{package}
\name{cloudview-package}
\alias{cloudview-package}
\alias{_PACKAGE}
\title{cloudview: Serves htmlwidgets over ssh}
\description{
Enables viewing htmlwidgets over ssh.
}
\author{
\strong{Maintainer}: Anthony North \email{anthony.jl.north@gmail.com}

Other contributors:
\itemize{
  \item Miles McBain \email{miles.mcbain@gmail.com} [contributor]
}

}
\keyword{internal}
