% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudview.R, R/enable.R
\name{cloudview}
\alias{cloudview}
\alias{enable_cloudview}
\alias{disable_cloudview}
\title{Cloudview}
\usage{
cloudview(x, ...)

enable_cloudview()

disable_cloudview()
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Serves htmlwidgets with \code{\link[servr:httd]{servr::httd()}}
}
\section{Functions}{
\itemize{
\item \code{enable_cloudview()}: Enable

\item \code{disable_cloudview()}: Disable

}}
\author{
Miles McBain
}
